hice la funcion para guardar nuevos productos

+++ src/main/java/com/example/sistema_ventas/modelo/clases/compra_producto.java
@@ -8,9 +8,9 @@ public class compra_producto {
 
     int id_producto;
 
-    int cant_producto;
+    double cant_producto;
 
-    public compra_producto(int id, int id_compra, int id_producto, int cant_producto) {
+    public compra_producto(int id, int id_compra, int id_producto, double cant_producto) {
         this.id = id;
         this.id_compra = id_compra;
         this.id_producto = id_producto;
@@ -41,11 +41,11 @@ public class compra_producto {
         this.id_producto = id_producto;
     }
 
-    public int getCant_producto() {
+    public double getCant_producto() {
         return cant_producto;
     }
 
-    public void setCant_producto(int cant_producto) {
+    public void setCant_producto(double cant_producto) {
         this.cant_producto = cant_producto;
     }
 }
